Add comments and fix spacing
