--- ios/FluentUI/Extensions/EnvironmentValues+Extensions.swift
@@ -1,24 +1,0 @@
-//
-//  Copyright (c) Microsoft Corporation. All rights reserved.
-//  Licensed under the MIT License.
-//
-import SwiftUI
-
-struct SafeAreaInsetsKey: EnvironmentKey {
-    static var defaultValue: EdgeInsets {
-        return UIApplication.shared.keyWindow?.safeAreaInsets.swiftUIInsets ?? EdgeInsets()
-    }
-}
-
-extension EnvironmentValues {
-  var swiftUIInsets: EdgeInsets {
-    get { self[SafeAreaInsetsKey.self] }
-    set { self[SafeAreaInsetsKey.self] = newValue }
-  }
-}
-
-private extension UIEdgeInsets {
-    var swiftUIInsets: EdgeInsets {
-        EdgeInsets(top: top, leading: left, bottom: bottom, trailing: right)
-    }
-}
